requirements.txt
langchain
fastapi
uvicorn
langserve
python-dotenv
openai
pandas
faiss-cpu
tiktoken